% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppCoalescenceModule.R
\docType{data}
\name{SimulationParameteres}
\alias{SimulationParameteres}
\alias{SimulationParameters}
\title{Container for the simulation parameters, to be edited and supplied to the SpatialTree class for simulation.}
\format{An object of class \code{list} of length 45.}
\usage{
SimulationParameters
}
\description{
Container for the simulation parameters, to be edited and supplied to the SpatialTree class for simulation.
}
\details{
Simulation parameter details
\itemize{
\item task the task reference number
\item seed the seed for setting random number generation
\item output_directory the path to the folder for storing simulation files
\item max_time the maximum number of seconds to simulate for before pausing
\item desired_specnum the desired number of species to aim for
\item times_list list of temporal sampling points
\item sigma mean dispersal distance for a normally-distributed dispersal kernel
\item dispersal_method the method of dispersal from "normal", "fat-tailed", "uniform" and "norm-uniform"
\item tau the tau parameter for a fat-tailed dispersal kernel
\item m_prob the probability of dispersing uniformly for the norm-uniform dispersal kernel
\item cutoff the maximum distance dispersal occurs for in a uniform dispersal kernel
\item dispersal_relative_cost the relative cost of moving through non-forest
\item restrict_self if true, prevents individuals from dispersing from their own cell
\item landscape_type type of landscape from "closed", "infinite" and "tiled"
\item dispersal_file a map of dispersal probabilities
\item reproduction_file a map of reproduction probabilies
\item fine_map_file fine resolution density map
\item coarse_map_file coarse resolution density map
\item sample_mask_file spatial sampling mask
\item grid_x_size x dimension of the grid
\item grid_y_size y dimension of the grid
\item sample_x_size x dimension of the sample mask
\item sample_y_size y dimension of the sample mask
\item sample_x_offset the x offset of the sample mask from the grid
\item sample_y_offset the y offset of the sample mask from the grid
\item fine_map_x_size the x dimension of the fine map
\item fine_map_y_size the y dimension of the fine map
\item fine_map_x_offset the x offset of the fine map from the sample mask
\item fine_map_y_offset the y offset of the fine map from the sample mask
\item coarse_map_x_size the x dimension of the coarse map
\item coarse_map_y_size the y dimension of the coarse map
\item coarse_map_x_offset the x offset of the coarse map from the fine map
\item coarse_map_y_offset the y offset of the coarse map from the fine map
\item coarse_map_scale the relative scale of the coarse map
\item deme the number of individuals per cell
\item deme_sample the global sampling proportion
\item uses_spatial_sampling if true, the sample mask defines relative sampling proportions across the map
\item pristine_fine_map the pristine fine map file
\item pristine_coarse_map the pristine coarse map file
\item gen_since_pristine the number of generations since the pristine state
\item habitat_change_rate the rate of change to the pristine map
\item speciation_rate the minimum speciation rate for the simulation
\item is_protracted if true, simulation will be simulated with protracted speciation
\item min_speciation_gen minimum number of generations required before speciation
\item max_speciation_gen maximum number of generations required before speciation
}
}
\keyword{datasets}
