% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppCoalescenceModule.R
\docType{data}
\name{SpeciationParameters}
\alias{SpeciationParameters}
\title{Container for speciation parameters, to be edited and supplied to the SpatialTree class
during post-simulation coalescence tree building.}
\format{An object of class \code{list} of length 10.}
\usage{
SpeciationParameters
}
\description{
Container for speciation parameters, to be edited and supplied to the SpatialTree class
during post-simulation coalescence tree building.
}
\details{
Speciation parameters details.
\itemize{
\item output_file the directory to output to, defaults to "none"
\item use_spatial if true, records full spatial locations of all individuals. Default=FALSE
\item sample_file supply a mask for defining spatial sampling. Default="null"
\item use_fragments supply a file containing fragment coordinates, or TRUE to let program calculate fragments
\item speciation_rates list of speciation rates to apply
\item times_list list of times to calculate coalescence tree for
\item min_speciation_gen the minimum number of generations required before speciation
\item max_speciation_gen the maximum number of generations required before speciation
\item metacommunity_size the number of individuals in the metacommunity
\item metacommunity_speciation_rate the effective speciation rate of the metacommunity
}
}
\keyword{datasets}
