% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppCoalescenceModule.R
\docType{class}
\name{Tree}
\alias{Tree}
\title{Non-spatial neutral simulations}
\description{
Run non-spatial neutral models. Here no map files are required, simply set up the
simulation and run. Dispersal parameters (sigma, tau, etc) are not relevant for this class.
}
\section{Methods}{

\describe{
\item{\code{applySpeciationRates(speciation_rates, output_file = "none",
  times_list = c(0), metacommunity_size = 0,
  metacommunity_speciation_rate = 0)}}{Applies the provided speciation parameters to the simulation}

\item{\code{checkOutputDatabase()}}{Checks that the output database exists}

\item{\code{getCommunityReferences()}}{Returns a data frame containing all the community references and their parameter sets.}

\item{\code{getMetacommunityReferences()}}{Returns a data frame containing all the metacommunity references and their parameter sets.}

\item{\code{getSpeciesAbundances(community_reference = NA)}}{Gets a data frame of species abundances where the community reference matches the input}

\item{\code{getSpeciesLocations(community_reference = 1)}}{Gets a data frame of species locations where the community reference matches the input.}

\item{\code{getSpeciesRichness(community_reference = NA)}}{Gets the community reference from the output database, or from the internal object if no 
community reference is supplied (this will return the last calculated species richness).}

\item{\code{output()}}{Outputs the biodiversity data to an sql database.}

\item{\code{setKeyParameters(task, seed, output_directory = "output", max_time = 3600,
  desired_specnum = 1, times_list = c(0), uses_logging = NA, deme = 1,
  deme_sample = 1)}}{Sets the key parameters for the simulation}

\item{\code{setOutputDatabase(output_db)}}{Checks the output database exists and has been created properly, if it has been set.}

\item{\code{setSimulationParameters(task, seed, speciation_rate,
  output_directory = "output", max_time = 3600, desired_specnum = 1,
  times_list = c(0), uses_logging = NA, deme = 1, deme_sample = 1)}}{Sets all simulation parameters}

\item{\code{setSpeciationParameters(speciation_rate)}}{Sets the speciation parameters for the simulation}
}}

\section{Alternative classes}{

For alternative classes providing similar functionality, see \link[=NeutralTree]{here}.
}

\section{Simulation parameters}{

These parameters are required at simulation time. The vast majority have defaults,
which are defined in setSimulationParameters()
\itemize{
\item \emph{task}: the task reference number
\item \emph{seed}: the seed for setting random number generation
\item \emph{output_directory}: the path to the folder for storing simulation files
\item \emph{max_time}: the maximum number of seconds to simulate for before pausing
\item \emph{desired_specnum}: the desired number of species to aim for
\item \emph{times_list}: list of temporal sampling points
\item \emph{uses_logging}: if true, all outputs are written to console
\item \emph{deme}: the number of individuals per cell
\item \emph{deme_sample}: the global sampling proportion
\item \emph{speciation_rate}: the minimum speciation rate for the simulation
}
}

\section{Post-simulation parameters}{

These are for rebuilding the coalescence tree under different conditions.
\itemize{
\item \emph{output_file}: the directory to output to, defaults to "none"
\item \emph{use_spatial}: if true, records full spatial locations of all individuals. Default=FALSE
\item \emph{sample_file}: supply a mask for defining spatial sampling. Default="null"
\item \emph{use_fragments}: supply a file containing fragment coordinates, or TRUE to let program calculate fragments
\item \emph{speciation_rates}: list of speciation rates to apply
\item \emph{times_list}: list of times to calculate coalescence tree for
\item \emph{min_speciation_gens}: list of the minimum number of generations required before speciation
\item \emph{max_speciation_gens}: list of the maximum number of generations required before speciation
\item \emph{metacommunity_size}: the number of individuals in the metacommunity
\item \emph{metacommunity_speciation_rate}: the effective speciation rate of the metacommunity
}
}

