% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppCoalescenceModule.R
\docType{package}
\name{rcoalescence}
\alias{rcoalescence}
\alias{rcoalescence-package}
\title{rcoalescence: Efficient spatial and non-spatial neutral ecology simulator written in C++}
\description{
A convenience wrapper for necsim, a Neutral Ecology Coalescence SIMulator for 
spatially-explicit neutral models based in c++.

rcoalescence allows for large-scale simulations to be performed on fully spatial files, with 
features to suit a wide variety of use cases.

Four classes exist to represent non-spatial/spatial and non-protracted/protracted simulations. 
See \link[=NeutralTree]{here} for a full description.
}
\section{Installation}{

There are several dependencies required before installation of this package will be successful.
\itemize{
    \item Install boost library for c++, available \href{https://www.boost.org/}{here}.
    \item Install gdal library for c++, available \href{http://www.gdal.org/}{here}.
    \item Install the sqlite3 library for c++, available \href{https://www.sqlite.org/index.html}{here}.
}
}

\section{Starting a simulation}{


Methods are contained within the relevant classes \link[=NeutralTree]{(see here)} for setting up,
running and analysing simulations. The general process is
\itemize{
    \item Create a new \link{Tree}, \link{SpatialTree}, \link{ProtractedTree} or
          \link{ProtractedSpatialTree} object, depending on use case.
    \item Set the simulation parameters and add additional historical maps.
    \item Run the simulation (generally the most time-consuming step).
    \item Apply speciation rates and other requirements post simulation using 
    SpeciationParameters.
    \item Get the species richness from the simulation.
    \item Optionally write results to an SQLite database for further manual analysis.
    \item If written to an output database, acquire the desired biodiversity metrics, such as 
    species locations or species abundances.

}
}

\author{
Sam Thompson
}
