% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppCoalescenceModule.R
\docType{class}
\name{TreeSimulation}
\alias{TreeSimulation}
\title{Non-spatial neutral TreeSimulations}
\description{
Run non-spatial neutral models. Here no map files are required, simply set up the
TreeSimulation and run. Dispersal parameters (sigma, tau, etc) are not relevant for this class.
}
\section{Methods}{

\describe{
\item{\code{applySpeciationRates(speciation_rates = NA, output_file = "none",
  times_list = c(0), metacommunity_option = c(), metacommunity_size = c(),
  metacommunity_speciation_rate = c(),
  metacommunity_external_reference = c())}}{Applies the provided speciation parameters to the TreeSimulation}

\item{\code{checkOutputDatabase()}}{Checks that the output database exists}

\item{\code{getCommunityReferences()}}{Returns a data frame containing all the community references and their parameter sets.}

\item{\code{getMetacommunityReferences()}}{Returns a data frame containing all the metacommunity references and their parameter sets.}

\item{\code{getSpeciesAbundances(community_reference = 1)}}{Gets a data frame of species abundances where the community reference matches the input}

\item{\code{getSpeciesLocations(community_reference = 1)}}{Gets a data frame of species locations where the community reference matches the input.}

\item{\code{getSpeciesRichness(community_reference = 1)}}{Gets the community reference from the output database, or from the internal object if no 
community reference is supplied (this will return the last calculated species richness).}

\item{\code{output()}}{Outputs the biodiversity data to an sql database.}

\item{\code{setOutputDatabase(output_db)}}{Checks the output database exists and has been created properly, if it has been set.}

\item{\code{setSimulationParameters(task, seed, min_speciation_rate,
  output_directory = "output", max_time = 3600, desired_specnum = 1,
  times_list = c(0), uses_logging = NA, deme = 1, deme_sample = 1)}}{Sets the key parameters for the TreeSimulation}

\item{\code{setTreeSimulationParameters(task, seed, speciation_rate,
  output_directory = "output", max_time = 3600, desired_specnum = 1,
  times_list = c(0), uses_logging = NA, deme = 1, deme_sample = 1)}}{Sets all TreeSimulation parameters}
}}

\section{Alternative classes}{

For alternative classes providing similar functionality, see \link[=NeutralTreeSimulation]{here}.
}

\section{TreeSimulation parameters}{

These parameters are required at TreeSimulation time. The vast majority have defaults,
which are defined in setTreeSimulationParameters()
\itemize{
\item \emph{task}: the task reference number
\item \emph{seed}: the seed for setting random number generation
\item \emph{output_directory}: the path to the folder for storing TreeSimulation files
\item \emph{max_time}: the maximum number of seconds to simulate for before pausing
\item \emph{desired_specnum}: the desired number of species to aim for
\item \emph{times_list}: list of temporal sampling points
\item \emph{uses_logging}: if true, all outputs are written to console
\item \emph{deme}: the number of individuals per cell
\item \emph{deme_sample}: the global sampling proportion
\item \emph{speciation_rate}: the minimum speciation rate for the TreeSimulation
}
}

\section{Post-TreeSimulation parameters}{

These are for rebuilding the coalescence TreeSimulation under different conditions.
\itemize{
\item \emph{output_file}: the directory to output to, defaults to "none"
\item \emph{use_spatial}: if true, records full spatial locations of all individuals. Default=FALSE
\item \emph{sample_file}: supply a mask for defining spatial sampling. Default="null"
\item \emph{use_fragments}: supply a file containing fragment coordinates, or TRUE to let program calculate fragments
\item \emph{speciation_rates}: list of speciation rates to apply
\item \emph{times_list}: list of times to calculate coalescence TreeSimulation for
\item \emph{min_speciation_gens}: list of the minimum number of generations required before speciation
\item \emph{max_speciation_gens}: list of the maximum number of generations required before speciation
\item \emph{metacommunity_size}: the number of individuals in the metacommunity
\item \emph{metacommunity_speciation_rate}: the effective speciation rate of the metacommunity
}
}

\examples{
# Example simulation for non-spatial models
library(rcoalescence)
simulation <- Tree$new()
# Simulate 100 individuals
simulation$setSimulationParameters(seed=10, task =1, output_directory = "output", 
                                   speciation_rate = 0.5, deme=100, uses_logging = TRUE)
simulation$runSimulation()

# Apply additional speciation rates
simulation$applySpeciationRates(speciation_rates = c(0.001, 0.7, 0.8))
# Output to a database
simulation$output()
# Get the species richness
simulation$getSpeciesRichness(1)
# Get the species abundances (last 5)
tail(simulation$getSpeciesAbundances(1), 5)

}
